requirements.txt
fastapi
uvicorn[standard]
celery[redis]
python-dotenv
modelscope
diffusers[torch]
torch
imageio
pydantic
python-multipart
redis
ffmpeg-python